# Quotation optimizations