I don't believe it's necessary to use a dictionary with design time property names, because we can rely on runtime column verification. Therefore, records can be erased to an array, providing an order of magnitude better runtime performance at a fraction of the memory footprint, and the quotations end up a lot simpler. It would be great if you could measure if that makes any difference during design time.

Furthermore, the `rowMapping` quotation could possibly be removed in the future completely, since it's based on result set definitions, which are passed to the runtime bit too