@kerams 2020-03-01
```fsharp
[<BenchmarkAttribute>]
member _.Before () =
    use cmd = DvdRental.CreateCommand<"select * from film">(dvdRental)
    cmd.Execute()
```


``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
AMD Ryzen 7 3700X, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT DEBUG
  DefaultJob : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT


```
| Method |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|----------:|----------:|------:|------:|------:|----------:|
| Before | 8.631 ms | 0.1713 ms | 0.3613 ms |     - |     - |     - |   6.06 MB |
| After | 7.938 ms | 0.0464 ms | 0.0362 ms |     - |     - |     - |   3.76 MB |

This is the runtime difference of selecting a 1000 rows before and after the optimization is applied. The speed is not really relevant as it largely depends on Postgres.

Wish I could benchmark the design time experience in VS, but I don't know how.