@kerams 2020-03-02
I've updated the benchmark having optimized array access further. We're now looking at a 20x speedup, which is what I had in mind originally.