@kerams 2020-03-02
Alright, I did just that. The tests pass, but a couple more sets of eyes would be nice.